Move the deliverable demo type input into a separate file, as there won't be an exposed resolver for this.
