--- server/src/model/deliverableDemonstrationType/deliverableDemonstrationTypeSchema.ts
@@ -1,16 +1,0 @@
-import { TagName } from "../../types";
-import { gql } from "graphql-tag";
-
-export const deliverableDemonstrationTypeSchema = gql`
-  input SetDeliverableDemonstrationTypesInput {
-    deliverableId: ID!
-    demonstrationId: ID!
-    demonstrationTypes: [TagName!]!
-  }
-`;
-
-export interface SetDeliverableDemonstrationTypesInput {
-  deliverableId: string;
-  demonstrationId: string;
-  demonstrationTypes: TagName[];
-}
