+++ server/src/model/deliverableDemonstrationType/deliverableDemonstrationTypeTypes.ts
@@ -1,0 +1,7 @@
+import { TagName } from "../../types";
+
+export type SetDeliverableDemonstrationTypesInput = {
+  deliverableId: string;
+  demonstrationId: string;
+  demonstrationTypes: TagName[];
+};
