+++ server/src/model/deliverableDemonstrationType/setDeliverableDemonstrationTypes.ts
@@ -1,5 +1,5 @@
 import { prisma, PrismaTransactionClient } from "../../prismaClient";
-import { SetDeliverableDemonstrationTypesInput } from "../../types";
+import { SetDeliverableDemonstrationTypesInput } from ".";
 import { Deliverable as PrismaDeliverable } from "@prisma/client";
 import { deleteAllDeliverableDemonstrationTypes, insertDeliverableDemonstrationTypes } from ".";
 import { getDeliverable } from "../deliverable";
