+++ server/src/types.ts
@@ -112,8 +112,6 @@ export type { Tag } from "./model/tag/tagSchema.js";
 
 export type { Deliverable, CreateDeliverableInput } from "./model/deliverable/deliverableSchema";
 
-export type { SetDeliverableDemonstrationTypesInput } from "./model/deliverableDemonstrationType/deliverableDemonstrationTypeSchema";
-
 export type DemonstrationTypeStatus = (typeof DEMONSTRATION_TYPE_STATUSES)[number];
 export type ClearanceLevel = (typeof CLEARANCE_LEVELS)[number];
 export type LocalDate = string & { readonly __brand: "LocalDate" };
